requirements.txt
Flask==2.3.3
Pillow==10.2.0
joblib==1.4.0
numpy==1.26.4
requests==2.31.0
tldextract==5.1.2
scikit-learn==1.4.1.post1
urllib3==2.2.1
gunicorn==21.2.0
